+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Add the `min_mask` argument to `ViralGeneFinder` to control the minimum lenght of masked regions on `mask=True`.
+
 ## [v0.1.0] - 2023-09-17
 [v0.1.0]: https://github.com/althonos/pyrodigal-gv/compare/13f7fb0...v0.1.0